placeクラスを廃止し、commandクラスにまとめる
